Rename files to match project style
