+++ tools/strict-null-checks/eligible-file-finder.js
@@ -1,6 +1,6 @@
 // @ts-check
 const path = require('path');
-const { getImportsForFile } = require('./tsHelper');
+const { getImportsForFile } = require('./import-finder');
 const glob = require('glob');
 const config = require('./config');
 
@@ -15,10 +15,14 @@ const forEachFileInSrc = (srcRoot, options) => {
                 return reject(err);
             }
 
-            return resolve(files.filter(file =>
-                !file.endsWith('.d.ts')
-                && (options && options.includeTests ? true : !/\.test\.tsx?$/.test(file))));
-        })
+            return resolve(
+                files.filter(
+                    file =>
+                        !file.endsWith('.d.ts') &&
+                        (options && options.includeTests ? true : !/\.test\.tsx?$/.test(file)),
+                ),
+            );
+        });
     });
 };
 module.exports.forEachFileInSrc = forEachFileInSrc;
@@ -42,7 +46,7 @@ module.exports.forStrictNullCheckEligibleFiles = async (repoRoot, forEach, optio
         const importList = getImportsForFile(file, srcRoot);
         imports.set(file, importList);
         return importList;
-    }
+    };
 
     const files = await forEachFileInSrc(srcRoot, options);
 
@@ -60,7 +64,10 @@ module.exports.forStrictNullCheckEligibleFiles = async (repoRoot, forEach, optio
                     }
                     // Don't treat cycles as blocking
                     const impImports = getMemoizedImportsForFile(imp, srcRoot);
-                    return impImports.filter(x => x !== file).filter(x => !checkedFiles.has(x)).length !== 0;
+                    return (
+                        impImports.filter(x => x !== file).filter(x => !checkedFiles.has(x))
+                            .length !== 0
+                    );
                 });
 
             const isEdge = nonCheckedImports.length === 0;
@@ -69,10 +76,12 @@ module.exports.forStrictNullCheckEligibleFiles = async (repoRoot, forEach, optio
             }
             return isEdge;
         });
-}
+};
 
 async function getCheckedFiles(tsconfigContent, tsconfigDir) {
-    const set = new Set(tsconfigContent.files.map(f => path.join(tsconfigDir, f).replace(/\\/g, '/')));
+    const set = new Set(
+        tsconfigContent.files.map(f => path.join(tsconfigDir, f).replace(/\\/g, '/')),
+    );
     const includes = tsconfigContent.include.map(include => {
         return new Promise((resolve, reject) => {
             glob(path.join(tsconfigDir, include), (err, files) => {
@@ -84,7 +93,7 @@ async function getCheckedFiles(tsconfigContent, tsconfigDir) {
                     set.add(file);
                 }
                 resolve();
-            })
+            });
         });
     });
     await Promise.all(includes);
